fix(pipe): widen metadata_tp dtype from uint8 to uint16
The extract step writes a `metadata_tp` (time point) column for every row
of the profiles parquet, casting the value through `pyarrow.uint8()`. Any
time point >=256 overflows the C uint8 type and aborts the parquet write:

    File "src/aliby/pipe.py", line 596, in get_profiles_from_state
        pyarrow.array([tp] * len(table), pyarrow.uint8()),
    pyarrow.lib.ArrowInvalid: Value 256 too large to fit in C integer type

Yeast time-lapses routinely exceed 256 time points (our Swain-lab corpus
clamps to 270 tps per zarr), so this failure mode is hit on every
sufficiently long run. 36 of 388 sites in the corpus failed on this in a
single batch.

Widen the column to `pyarrow.uint16()` (max 65535) which is still narrow
enough for parquet efficiency but covers any realistic time-lapse length.
No behavioural change for runs with <=255 time points: uint16 is a
superset of uint8.

+++ src/aliby/pipe.py
@@ -593,7 +593,7 @@ def get_profiles_from_state(state: dict, pipeline: dict) -> pyarrow.Table:
                 )
                 table = table.append_column(
                     "metadata_tp",
-                    pyarrow.array([tp] * len(table), pyarrow.uint8()),
+                    pyarrow.array([tp] * len(table), pyarrow.uint16()),
                 )
                 data[step_prefix].append(table)
